% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_centering.R
\name{d_centering}
\alias{d_centering}
\title{Within-individual centering function}
\usage{
d_centering(data, centering_by, cluster = NULL, variable)
}
\arguments{
\item{data}{data frame}

\item{centering_by}{individual ID}

\item{cluster}{higher level structure. E.g. levels in centering_by further structured in cluster. NULL as default}

\item{variable}{character name or vector of characters. Variables to be partitioned.If more than one,please, concatenate using 'c'.}
}
\value{
New variables added to the end of the original data frame with 'mean_' prefix for the mean value of variable for every centering_by level and prefix 'cent' for the within-indivual term.
}
\description{
Partition of variation within- and among-individuals following the method summarised by Van de Pol & Wright (2009) (doi: 10.1016/j.anbehav.2008.11.006)
}
