% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelInt_predict.R
\name{modelInt_predict}
\alias{modelInt_predict}
\title{Prediction interval for a fitted \code{merMod} object. See \code{\link{lmer}}, See \code{\link{glmer}}.}
\usage{
modelInt_predict(model, df, n_sim = 100, prob = 0.95, cluster = NULL,
  cluster_name = NULL, ...)
}
\arguments{
\item{model}{fitted \code{merMod} object}

\item{df}{data frame with new combination of fixed effect values to generate prediction}

\item{n_sim}{number of bootstrap simulation}

\item{prob}{probability for the predicted interval}

\item{cluster}{logical. TRUE / FALSE for whether to use parallele computing. See \code{\link{bootMer}}}

\item{cluster_name}{if cluster = TRUE, a user-defined cluster name}

\item{...}{see \code{\link{bootMer}} for more argument options}
}
\value{
Data frame with prediction interval
}
\description{
This function uses \code{\link{bootMer}} to bootstrap a mixed model and calculate
prediction intervals.
}
\details{
See \code{\link{bootMer}} and \code{\link{boot}} for further details.
information about the presence/absence of the incubating individual in the nest.
}
