% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wbsw_gc.R
\name{wbsw_gc}
\alias{wbsw_gc}
\title{Calculate group composition using, the most famous, Careers table.}
\usage{
wbsw_gc(Careers, career_startDate, career_endDate, career_birdID,
  career_season, season, group, interval, all_birds, maturationTime,
  time_steps)
}
\arguments{
\item{Careers}{name of the sparrow weaver project careers table loaded in R}

\item{career_startDate}{character type. Name of the column specifying residence start times in Careers table, for GC calculations}

\item{career_endDate}{character type. Name of the column specifying residence end times in Careers table, for GC calculations}

\item{career_birdID}{character type. Name of the column specifying bird ID in Careers table. Please, note that this name has to be common across the Careers and All Birds tables}

\item{career_season}{character type. Name of the column specifying season in Careers table}

\item{season}{for a specific time interval, define the season}

\item{group}{group name for group composition calculation}

\item{interval}{vector of length 2. Each element follwing format = Y-m-d. The first element corresponds to the start date of the interval}

\item{all_birds}{name of the All Birds table loaded in R}

\item{maturationTime}{Number of days a bird is considered as a fledgling after hatching. It serves to count birds inthe first days of life}

\item{time_steps}{number of intervals to divide period of gc calculations}
}
\value{
A vector with the following information
}
\description{
This function takes time interval and calculates group composition.
}
